Add friends likes - Спринт 9

+++ src/main/java/ru/yandex/practicum/filmorate/util/IdGen.java
@@ -3,11 +3,15 @@ package ru.yandex.practicum.filmorate.util;
 import lombok.Getter;
 import lombok.Setter;
 
+/**
+ * Генератор ID
+ */
 public class IdGen {
     @Getter
     @Setter
-    private int id = 0;
-    public int nextId() {
+    private long id = 0;
+
+    public long nextId() {
         return ++id;
     }
 }
